Use the env script we use on hod connect to set env for hod batch.
This fixes a bug where hod used the local job runner which gives the appearance
of working; but was absolutely not what we wanted.
